+++ utils-screeen-base.go
@@ -15,7 +15,7 @@ var rigorousSec = 1
 //rigorous : 是否需要持續1秒鐘以上圖片存在才算匹配成功 秒數可由rigorousSec更改,
 //imgFullPaths : 傳入圖片相對全路徑 (多張),
 //回傳 : 找到與否 - 找到的是第幾張 - 找到的中心點x - 找到的中心點y
-func findOneImages(frequency int, matchNumber float32, rigorous bool, imgFullPaths ...string) (succ bool, index int, x int, y int) {
+func findOneImages(frequency int, matchNumber float64, rigorous bool, imgFullPaths ...string) (succ bool, index int, x int, y int) {
 	log.Infof("findOneImages, frequency: %d, matchNumber: %e, rigorous: %s", frequency, matchNumber, rigorous)
 	succ = false
 	index = -1
@@ -74,7 +74,7 @@ func findOneImages(frequency int, matchNumber float32, rigorous bool, imgFullPat
 //rigorous : 是否需要持續1秒鐘以上圖片存在才算匹配成功 : 秒數可由rigorousSec修改,
 //imgFullPaths : 傳入圖片相對全路徑 (多張),
 //回傳 : 找到與否
-func findAllImages(frequency int, matchNumber float32, rigorous bool, imgFullPaths ...string) (succ bool) {
+func findAllImages(frequency int, matchNumber float64, rigorous bool, imgFullPaths ...string) (succ bool) {
 	log.Infof("findAllImages, frequency: %d, matchNumber: %e, rigorous: %s", frequency, matchNumber, rigorous)
 	succ = false
